successfully merged structure and tsx
