+++ screens/PostingScreen.tsx
@@ -2,8 +2,6 @@ import { View, Text, StyleSheet } from "react-native";
 
 
 function PostingScreen(){
-  
-
   return(
     <View style={styles.container}>
       <Text>Posting Screen</Text>
